requirements.txt
pluginmanager
PyAudio
pyzmq
